# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rezidence4 API",
    "version": "1.0.0",
    "description": "DomHub multi-tenant API. Platform-v1 routes под /api/v1/* serve as source of truth (см. CLAUDE.md). Этот spec — минимальный contract anchor для openapi_route.test.js + клиентов; полные request/response schemas наполняются инкрементально (см. docs/api/README.md)."
  },
...
  "paths": {
...
    "/api/v1/chat/messages": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessagesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "ChatMessage": {
        "type": "object"
      },
      "ChatMessagesResponse": {
        "type": "object",
        "required": [
          "messages",
          "hasMore"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}